requirements.txt
django~=5.1.4
djangorestframework~=3.15.2
rest_framework_simplejwt>=5.3.1